requirements.txt
appdirs==1.4.3
cachetools==3.1.0
certifi==2018.11.29
chardet==3.0.4
google-api-core==1.8.0
google-auth==1.6.3
google-cloud-bigquery==1.9.0
google-cloud-core==0.29.1
google-cloud-language==1.1.1
google-cloud-speech==0.36.3
google-cloud-storage==1.14.0
google-cloud-texttospeech==0.4.0
google-resumable-media==0.3.2
googleapis-common-protos==1.5.8
grpcio==1.18.0
idna==2.8
mypy==0.670
mypy-extensions==0.4.1
protobuf==3.6.1
pyasn1==0.4.5
pyasn1-modules==0.2.4
pytz==2018.9
requests==2.21.0
rsa==4.0
six==1.12.0
typed-ast==1.3.1
urllib3==1.24.1
web-cache==1.1.0
